requirements.txt
celery==5.2.7
flask==2.2.3
flask_cors==3.0.10
flask-jwt-extended==4.4.4
flask_restful==0.3.9
flask_marshmallow==0.15.0
flask_sqlalchemy==3.0.3
google-cloud-storage==2.8.0
gunicorn==20.1.0
gevent==22.10.2
marshmallow_sqlalchemy==0.29.0
psycopg2==2.9.6
redis==4.5.4